requirements.txt
# ─────────────────────────────────────────────
# Interface do usuário
# ─────────────────────────────────────────────
streamlit==1.32.0           # Framework do app web

# ─────────────────────────────────────────────
# Manipulação de dados
# ─────────────────────────────────────────────
pandas==2.2.1               # Manipulação de tabelas/dataframes
numpy==1.26.4               # Operações numéricas

# ─────────────────────────────────────────────
# Machine Learning (modelos tabulares)
# ─────────────────────────────────────────────
scikit-learn==1.4.1.post1   # Random Forest, métricas, pré-processamento
                             # post1 = versão compatível com Python 3.12
xgboost==2.0.3              # Modelo XGBoost (gradient boosting)

# ─────────────────────────────────────────────
# Explicabilidade dos modelos
# ─────────────────────────────────────────────
shap==0.45.0               # SHAP values para explicar predições

# ─────────────────────────────────────────────
# Deep Learning (CNN para imagens)
# ─────────────────────────────────────────────
tensorflow==2.17.0         # Framework de deep learning (inclui Keras)

# ─────────────────────────────────────────────
# Processamento de imagens
# ─────────────────────────────────────────────
Pillow==10.2.0              # Leitura e manipulação de imagens
opencv-python-headless==4.9.0.80  # Processamento de imagem (Grad-CAM)
                                   # versão headless = sem interface gráfica
                                   # (necessário pra rodar em Docker/servidor)

# ─────────────────────────────────────────────
# Visualizações
# ─────────────────────────────────────────────
matplotlib==3.8.3           # Gráficos gerais e Grad-CAM overlay
seaborn==0.13.2             # Gráficos estatísticos (heatmaps, distribuições)
plotly==5.20.0              # Gráficos interativos no Streamlit

# ─────────────────────────────────────────────
# Utilitários
# ─────────────────────────────────────────────
joblib==1.3.2               # Salvar/carregar modelos .pkl
